+++ CPAC/pipeline/nipype_pipeline_engine/engine.py
@@ -3,10 +3,39 @@ See https://fcp-indi.github.io/docs/developer/nodes
 for C-PAC-specific documentation.
 See https://nipype.readthedocs.io/en/latest/api/generated/nipype.pipeline.engine.html
 for Nipype's documentation.
-'''  # noqa: E501
+
+STATEMENT OF CHANGES:
+    This file is derived from sources licensed under the Apache-2.0 terms,
+    and this file has been changed.
+
+CHANGES:
+    * Supports just-in-time dynamic memory allocation
+    * Skips doctests that require files that we haven't copied over
+    * Applies a random seed
+    * Supports overriding memory estimates via a log file and a buffer
+
+ORIGINAL WORK'S ATTRIBUTION NOTICE:
+    Copyright (c) 2009-2016, Nipype developers
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+        http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+
+    Prior to release 0.12, Nipype was licensed under a BSD license.
+
+Modifications Copyright (C) 2022  C-PAC Developers
+
+This file is part of C-PAC.'''  # noqa: E501
 import os
 import re
-import warnings
 from logging import getLogger
 from inspect import Parameter, Signature, signature
 from nibabel import load